+++ changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - bleeding effect
 - burning effect
 - poisoned effect
+- concentrating effect
 
 ### Fixed